+++ changelog.md
@@ -3,6 +3,10 @@
 - Fixed Flint and Steel being usable to place fire in protected Areas
 ## Settings & Config
 ## Commands
+- Fixed `/ultracraft progression level grant-all` not unlocking Level `clair de lune`
+- Fixed `/ultracraft progression <unlocked//obtained> grant-all` not unlocking `Hivel Wings`, `Blood Healing` and `UltraHUD`
+- Added `/ultracraft progression <list> revoke-all <target>`
+  - Clears the given progression list on given targets (level unlocks are still global, so any target will affect everyone)
 ## Tweaks
 ## Minor Changes
 - Fixed Shotgun Pump Animation